商品詳細画面まで遷移できるようになりました。

+++ src/main/java/jp/ken/jdbc/domain/entity/GenreEntity.java
@@ -2,13 +2,13 @@ package jp.ken.jdbc.domain.entity;
 
 public class GenreEntity {
 
-    private int genreId;
+    private long genreId;
     private String genreName;
 
-    public int getGenreId() {
+    public long getGenreId() {
         return genreId;
     }
-    public void setGenreId(int genreId) {
+    public void setGenreId(long genreId) {
         this.genreId = genreId;
     }
